pyproject.toml
[project]
name = "confiture"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.4",
    "pytest-django>=4.9.0",
    "pytest>=8.3.4",
    "django-extensions>=3.2.3",
    "django-cotton>=1.5.1",
    "django-magic-context",
    "attrs>=24.3.0",
    "django-template-partials>=24.4",
    "django-lucide-icons>=0.2.2",
    "neapolitan>=24.8",
    "django-allauth>=65.3.0",
    "django-anymail[brevo]>=12.0",
    "sqids>=0.5.0",
    "werkzeug>=3.1.3",
    "pluggy>=1.5.0",
    "httpx>=0.28.1",
    "django-harlequin>=1.4.0",
    "harlequin[postgres]>=1.25.2",
]

[dependency-groups]
dev = [
    "ruff>=0.8.4",
]

[tool.uv.sources]
django-magic-context = { git = "https://github.com/fbinz/django-magic-context" }

[tool.ruff]
lint.extend-select = ["I"]
exclude = ["**migrations**"]
